app_backend/requirements.txt
fastapi
uvicorn
pymongo
xmltodict
yt-dlp
requests
python-dotenv
